added test cases for require() and friends
   https://issues.apache.org/jira/browse/CB-267

Added the commonjs test suite from:

   https://github.com/commonjs/commonjs

Those tests are now in

   thirdparty/commonjs-tests

Test scaffolding has been added to

   test/commonjs

See the README.txt file in that directory for
instructions on running the tests.

In addition, the LICENSE was modified to include
information on the commonjs tests and existing
jasmine usage.  A NOTICE file was also added,
since it was not available before.