[CB-280] Improve layout of cordova-js scripts
https://issues.apache.org/jira/browse/CB-280

This was basically a refactoring of the lib subdirectories,
without changing any of the code IN the lib subdirectories.

Other changes:

- build/packager.js

modified to build based on the new lib subdirectories

- Jakefile

modified to call a new method in build/packager

- README.md

updated file/directory locations

- test/runner.js

changed the runner to run off a new 'test' platform,
instead of using an on-the-fly platform it pieced
together when the tests run.

also changed the static() URL bits for the
browser-based test - it was pulling in more
than it needed

- test/suite.html

changed the urls of the jasmine files, based
on the updates to the static() URL bits in the
runner.js
-
- lib/common/exec.js
- lib/common/platform.js

stub files that will be replaced with platform-specific
ones