[CB-762] Change BlackBerry plugin from Contact to Contacts.
The BlackBerry override of the common Contacts API expects the associated native plugin to be named "Contact" in plugins.xml.This differs from the common cordova-js expectation of "Contacts".This bug fix will change the BlackBerry implementation to match
common.

Existing Cordova BlackBerry projects will need updated to accountfor this change. Modify the plugins.xml file of the existingproject to change the following line:

from:
<plugin name="Contact" value="org.apache.cordova.pim.Contact"/>

to:
<plugin name="Contacts" value="org.apache.cordova.pim.Contact"/>