CB-1585 - simpler impl of iOS console bits
https://issues.apache.org/jira/browse/CB-1585
console.debug, console.info are undefined

The previous version of this code was getting busy,
and the fix to add debug/info were going to make
it even busier, so I made it simpler instead.

We now return the original console, if it exists,
and just patch the relevant methods on there.