[all] Use type-coercion for booleans in Camera
The reverts a change to argscheck that converts "true" to true. This
unsafe if the parameter is not actually a boolean, but instead a string.

This also changes the behaviour of negative numbers that were passed to
the boolean Camera arguments. Negative numbers will no longer be treated
as false.

Related to https://issues.apache.org/jira/browse/CB-2165