Drop build/test support for Node.js 6 (#190)
Some npm packages, including some of our build/test dependencies have
begun dropping support for Node.js 6 in their latest versions. To
facilitate dependency updates, we remove support for Node.js 6 here.

This package is usually not built and tested on our user's machines,
but is instead bundled with the platforms by us. Thus, this change
should not be considered a breaking change.