Stable error handling for loaded Cordova plugins
Platforms affected
All

Motivation and Context
When developing a large Cordova app in a distributed team occasionally a developer commits a Cordova plugin which contains an issue which directly results in a javaScript exception on plugin load. 
Before the fix the Cordova startup crashes and the app results in a whitescreen.

Description
The fix makes handlePluginsObject more stable. The function keeps track of the loaded modules. If a load error occurs for a module the module is removed from the modules list. Additionally an alert is displayed. This makes it way easier for developers to track and fix the error.